pyproject.toml
[tool.poetry]
name = "auth-system-project"
version = "0.1.0"
description = ""
authors = ["Henkamoh00 <henkamoh@gmail.com>"]
readme = "README.md"



[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.115.12"
uvicorn = "^0.34.3"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.41"}
dotenv = "^0.9.9"
passlib = "^1.7.4"
pydantic-settings = "^2.9.1"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
asyncpg = "^0.30.0"
pytest = "^8.4.0"
httpx = "^0.28.1"
pydantic = {extras = ["email"], version = "^2.11.5"}
python-multipart = "^0.0.20"
redis = "^6.2.0"
slowapi = "^0.1.9"
async-lru = "^2.0.5"
upstash-redis = "^1.4.0"
jinja2 = "^3.1.6"
itsdangerous = "^2.2.0"



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
